api/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
python-dotenv==0.21.0
requests==2.28.1

